surface boundary mod

+++ src/surface/distToPoint.cpp
@@ -2,15 +2,15 @@
 
 float NIBR::Surface::squaredDistToPoint(float *p) {
 
-    Eigen::MatrixXf point(1,3);
+    Eigen::MatrixXd point(1,3);
 
     point(0,0) = p[0];
     point(0,1) = p[1];
     point(0,2) = p[2];
 
-    Eigen::VectorXf sqrD;
+    Eigen::VectorXd sqrD;
     Eigen::VectorXi I;
-    Eigen::MatrixXf C;
+    Eigen::MatrixXd C;
     AABB_tree.squared_distance(V,F,point,sqrD,I,C);
 
     return sqrD(0);
@@ -25,15 +25,15 @@ float NIBR::Surface::distToPoint(float *p) {
         return NAN;
     }
 
-    Eigen::MatrixXf point(1,3);
+    Eigen::MatrixXd point(1,3);
 
     point(0,0) = p[0];
     point(0,1) = p[1];
     point(0,2) = p[2];
 
-    Eigen::VectorXf sqrD;
+    Eigen::VectorXd sqrD;
     Eigen::VectorXi I;
-    Eigen::MatrixXf C;
+    Eigen::MatrixXd C;
     AABB_tree.squared_distance(V,F,point,sqrD,I,C);
 
     float dist = isPointInside(p) ? std::sqrt(sqrD(0)) : -std::sqrt(sqrD(0));
@@ -45,15 +45,15 @@ float NIBR::Surface::distToPoint(float *p) {
 
 float NIBR::Surface::squaredDistToPoint(float *p, int& faceInd) {
 
-    Eigen::MatrixXf point(1,3);
+    Eigen::MatrixXd point(1,3);
 
     point(0,0) = p[0];
     point(0,1) = p[1];
     point(0,2) = p[2];
 
-    Eigen::VectorXf sqrD;
+    Eigen::VectorXd sqrD;
     Eigen::VectorXi I;
-    Eigen::MatrixXf C;
+    Eigen::MatrixXd C;
     AABB_tree.squared_distance(V,F,point,sqrD,I,C);
 
     faceInd = I(0);
@@ -69,15 +69,15 @@ float NIBR::Surface::distToPoint(float *p, int& faceInd) {
         return NAN;
     }
 
-    Eigen::MatrixXf point(1,3);
+    Eigen::MatrixXd point(1,3);
 
     point(0,0) = p[0];
     point(0,1) = p[1];
     point(0,2) = p[2];
 
-    Eigen::VectorXf sqrD;
+    Eigen::VectorXd sqrD;
     Eigen::VectorXi I;
-    Eigen::MatrixXf C;
+    Eigen::MatrixXd C;
     AABB_tree.squared_distance(V,F,point,sqrD,I,C);
 
     float dist = isPointInside(p) ? std::sqrt(sqrD(0)) : -std::sqrt(sqrD(0));
@@ -89,15 +89,15 @@ float NIBR::Surface::distToPoint(float *p, int& faceInd) {
 
 float NIBR::Surface::squaredDistToPoint(float *p, int& faceInd, float* closestPoint) {
 
-    Eigen::MatrixXf point(1,3);
+    Eigen::MatrixXd point(1,3);
 
     point(0,0) = p[0];
     point(0,1) = p[1];
     point(0,2) = p[2];
 
-    Eigen::VectorXf sqrD;
+    Eigen::VectorXd sqrD;
     Eigen::VectorXi I;
-    Eigen::MatrixXf C;
+    Eigen::MatrixXd C;
     AABB_tree.squared_distance(V,F,point,sqrD,I,C);
 
     closestPoint[0] = C(0,0);
@@ -117,15 +117,15 @@ float NIBR::Surface::distToPoint(float *p, int& faceInd, float* closestPoint) {
         return NAN;
     }
 
-    Eigen::MatrixXf point(1,3);
+    Eigen::MatrixXd point(1,3);
 
     point(0,0) = p[0];
     point(0,1) = p[1];
     point(0,2) = p[2];
 
-    Eigen::VectorXf sqrD;
+    Eigen::VectorXd sqrD;
     Eigen::VectorXi I;
-    Eigen::MatrixXf C;
+    Eigen::MatrixXd C;
     AABB_tree.squared_distance(V,F,point,sqrD,I,C);
 
     float dist = isPointInside(p) ? std::sqrt(sqrD(0)) : -std::sqrt(sqrD(0));
